Many changes introduced
Substitute wj into s and calculate frequency response for analog
filter. Do the same for phase. Analog magnitude works, phase the same.
Digital Filter does not give good results for magnitudes yet.
Introduced overlay plotting and couple helper functions.

+++ utils.c
@@ -12,7 +12,7 @@ size_t max_size_t(size_t a, size_t b)
     return (a > b) ? a : b;
 }
 
-void fill_n_with_step(float *xs, size_t size, float start, float stop)
+void fill_n_with_step(double *xs, size_t size, float start, float stop)
 {
     float step = (stop - start) / (float)(size);
     xs[0] = start;
